+++ mockaroo/__init__.py
@@ -1,6 +1,3 @@
-"""A Python Package for the Mockaroo API."""
-from .api.client import Client
-from .api.exceptions import MockarooError, InvalidApiKeyError, UsageLimitExceededError
+from .client import Client
 
-__all__ = ["Client", "MockarooError", "InvalidApiKeyError", "UsageLimitExceededError"]
-__version__ = "2.0.0"
+__all__ = ["Client"]