+++ mockaroo/api/constants.py
@@ -1,9 +1,0 @@
-"""Constants for interacting with the Mockaroo API."""
-
-TYPE_ENDPOINT = "/api/types/"
-UPLOAD_ENDPOINT = "/api/datasets/"
-GENERATE_ENDPOINT = "/api/generate"
-
-HTTP_GET = "GET"
-HTTP_POST = "POST"
-HTTP_DELETE = "DELETE"