+++ tests/test_client.py
@@ -1,14 +1,14 @@
 # ruff: noqa: D103
 """Unit tests for the mockaroo api module."""
-import pytest
+
 import os
 import warnings
+from unittest.mock import Mock, patch
 
-from unittest.mock import patch, Mock
+import pytest
 
-from mockaroo.api.client import Client
-from mockaroo.api.exceptions import InvalidApiKeyError, UsageLimitExceededError
-from mockaroo.api.constants import GENERATE_ENDPOINT, UPLOAD_ENDPOINT, TYPE_ENDPOINT
+from mockaroo import Client
+from mockaroo.exceptions import InvalidApiKeyError, UsageLimitExceededError
 
 
 def empty_client():
@@ -32,7 +32,7 @@ def test_no_api_key():
         assert len(w) == 1
         assert issubclass(w[-1].category, UserWarning)
         assert (
-            "API key is not provided. Set API_KEY `export API_KEY=your_api_key`."
+            "API key is not provided. Set MOCKAROO_API_KEY `export MOCKAROO_API_KEY=your_api_key`."
             == str(w[-1].message)
         )
 
@@ -108,61 +108,51 @@ def test_delete():
         assert result == {"success": True}
 
 
-def test_default_generate_with_schema():
+@pytest.mark.parametrize(
+    "fmt, expected_output",
+    [
+        (
+            "json",
+            [{"mock_key": "mock_value"}],
+        ),
+        ("csv", b"id,col_header1,col_header2\n1,row_value1,row_value2\n"),
+        ("txt", b"id\tcol_header1\tlcol_header2\n1\trow_value1\trow_value2\n"),
+        (
+            "xml",
+            b"<?xml version='1.0' encoding='UTF-8'?>\n<dataset>\n<record><id>1</id><first_name>Ericka</first_name><last_name>Stobart</last_name></record></dataset>",  # noqa: E501
+        ),
+    ],
+)
+def test_generate(fmt, expected_output):
     client = empty_client()
     with patch("requests.request") as mock_request:
         mock_response = Mock()
-        mock_response.json.return_value = [{"mock_key": "mock_value"}]
+        if fmt == "json":
+            mock_response.json.return_value = expected_output
+        else:
+            mock_response.content = expected_output
         mock_response.status_code = 200
         mock_request.return_value = mock_response
 
-        result = client.generate(schema="Person")
+        result = client.generate(schema="test_schema", fmt=fmt)
 
-        assert result == [{"mock_key": "mock_value"}]
+        assert result == expected_output
 
-
-def test_generate_format_csv():
-    client = empty_client()
-    with patch("requests.request") as mock_request:
-        mock_response = Mock()
-        mock_response.content = b"id,col_header1,col_header2\n1,row_value1,row_value2\n"
-        mock_response.status_code = 200
-        mock_request.return_value = mock_response
-
-        result = client.generate(schema="Person", fmt="csv")
-
-        assert result == b"id,col_header1,col_header2\n1,row_value1,row_value2\n"
-
-
-def test_generate_format_txt():
-    """Test generate for txt format."""
-    client = empty_client()
-    with patch("requests.request") as mock_request:
-        mock_response = Mock()
-        mock_response.content = (
-            b"id\tcol_header1\tlcol_header2\n1\trow_value1\trow_value2\n"
-        )
-        mock_response.status_code = 200
-        mock_request.return_value = mock_response
-
-        result = client.generate(schema="Person", fmt="csv")
-
-        assert result == b"id\tcol_header1\tlcol_header2\n1\trow_value1\trow_value2\n"
 
 def test_schema_and_fields_argument_warning():
     """Test raise `UserWarning`when both `schema` and `fields` parameters are given.
 
-    The purpose of this test is to verify that a warning is generated when a user 
-    specifies both schema and fields. The schema parameter takes precedence, meaning 
-    any fields specified will be ignored. The warning informs the user of this behavior 
+    The purpose of this test is to verify that a warning is generated when a user
+    specifies both schema and fields. The schema parameter takes precedence, meaning
+    any fields specified will be ignored. The warning informs the user of this behavior
     to avoid confusion.
 
     Example case:
 
         - If the `schema` is set to "Person", and
         - The user also specifies `fields` as [{"name", "id", "type", "Row Number"}],
-        
-        The test should confirm that a UserWarning is issued, and the generated data 
+
+        The test should confirm that a UserWarning is issued, and the generated data
         will only follow the "Person" schema, ignoring the custom fields.
 
     """
@@ -171,61 +161,48 @@ def test_schema_and_fields_argument_warning():
         record=True
     ) as w:
         mock_response = Mock()
-        mock_response.json.return_value = [{
-            "first_name": "Jane", 
-            "last_name": "Doe"
-        }]
+        mock_response.json.return_value = [{"first_name": "Jane", "last_name": "Doe"}]
         mock_response.status_code = 200
         mock_request.return_value = mock_response
 
         result = client.generate(
-            schema="Person", 
-            fields=[
-                {
-                    "name": "id", 
-                    "type": "Row Number"
-                }
-            ]
+            schema="Person", fields=[{"name": "id", "type": "Row Number"}]
         )
-        assert result == [{
-            "first_name": "Jane", 
-            "last_name": "Doe"
-        }]
+        assert result == [{"first_name": "Jane", "last_name": "Doe"}]
         assert len(w) == 1
         assert issubclass(w[-1].category, UserWarning)
         assert (
-            "You should specify either 'schema' or 'fields', "\
-            "but not both. `schema` will override any values passed"\
-            "to `fields`."
-            == str(w[-1].message)
+            "You should specify either 'schema' or 'fields', "
+            "but not both. `schema` will override any values passed"
+            "to `fields`." == str(w[-1].message)
         )
-    
+
 
 @pytest.mark.parametrize(
     "endpoint, params, expected_url",
     [
         (
-            GENERATE_ENDPOINT,
+            Client.GENERATE_ENDPOINT,
             {"schema": "Person", "fmt": "json", "count": 1},
             "https://api.mockaroo.com/api/generate.json?schema=Person&count=1&client=python&key=api_key",
         ),
         (
-            GENERATE_ENDPOINT,
+            Client.GENERATE_ENDPOINT,
             {"schema": "Person", "fmt": "csv", "count": 1},
             "https://api.mockaroo.com/api/generate.csv?schema=Person&count=1&client=python&key=api_key",
         ),
         (
-            GENERATE_ENDPOINT,
+            Client.GENERATE_ENDPOINT,
             {"schema": "Person", "fmt": "csv", "count": 5},
             "https://api.mockaroo.com/api/generate.csv?schema=Person&count=5&client=python&key=api_key",
         ),
         (
-            TYPE_ENDPOINT,
+            Client.TYPE_ENDPOINT,
             {},
             "https://api.mockaroo.com/api/types/?client=python&key=api_key",
         ),
         (
-            UPLOAD_ENDPOINT,
+            Client.UPLOAD_ENDPOINT,
             {"name": "test_name"},
             "https://api.mockaroo.com/api/datasets/test_name?client=python&key=api_key",
         ),