+++ tests/test_layout.py
@@ -1,8 +1,10 @@
 # ruff: noqa: D103
 """Unit tests for layout module."""
+
 import pytest  # noqa: F401
 
-from mockaroo.cli.layout import print_table
+from mockaroo.layout import print_table
+
 
 def test_pager_layout():
     data_types = [{"name": "Words", "type": "string", "parameters": None}]